kalosm_language_model

Trait IntoChatMessage

pub trait IntoChatMessage {
    // Required method
    fn into_chat_message(self) -> ChatMessage;
}

A trait for types that can be converted into a chat message.

Example

// Displayable types are converted into a user chat message
let user_message = "Hello, world!";
let chat_message = user_message.into_chat_message();
assert_eq!(chat_message.role(), MessageType::UserMessage);

// Or you can create a chat message manually
let chat_message = ChatMessage::new(MessageType::ModelAnswer, "Hello, world!".to_string());
assert_eq!(chat_message.role(), MessageType::ModelAnswer);

Required Methods

fn into_chat_message(self) -> ChatMessage

Convert the type into a chat message.