kalosm_language_model

Struct GenerationParameters

pub struct GenerationParameters { /* private fields */ }

Parameters to use when generating text.

Implementations

impl GenerationParameters

pub const fn new() -> Self

Create a new GenerationParameters

pub fn sampler(&self) -> SamplerChain

Create a sampler chain from the generation parameters.

pub fn with_top_p(self, top_p: f64) -> Self

Set the top_p parameter to the generation parameters (only used by the OpenAI API).

pub fn with_top_k(self, top_k: u32) -> Self

Set the top_k parameter to the generation parameters (only used by the Anthropic API).

pub fn mirostat2_sampler(self) -> SampleMirostat2

Get the mirostat2 sampler from the generation parameters.

pub fn bias_only_sampler(self) -> SamplerChain

Create a sampler chain from the generation parameters without removing any tokens. This can be useful in combination with [ModelExt::stream_structured_text_with_sampler] which may pick unlikely tokens.

pub fn with_temperature(self, temperature: f32) -> Self

Set the temperature to use when generating text.

pub fn with_tau(self, tau: f32) -> Self

Set the tau to use when generating text.

pub fn with_eta(self, eta: f32) -> Self

Set the eta to use when generating text.

pub fn with_mu(self, mu: f32) -> Self

Set the mu to use when generating text.

pub fn with_repetition_penalty(self, repetition_penalty: f32) -> Self

Set the repetition penalty to use when generating text.

pub fn with_repetition_penalty_range( self, repetition_penalty_range: u32, ) -> Self

Set the repetition penalty range to use when generating text.

pub fn with_max_length(self, max_length: u32) -> Self

Set the maximum length to use when generating text.

pub fn with_stop_on(self, stop_on: impl Into<Option<String>>) -> Self

Set the string to stop on when generating text.

pub fn with_seed(self, seed: impl Into<Option<u64>>) -> Self

Set the seed to use when generating text.

pub fn temperature(&self) -> f32

Get the temperature to use when generating text.

pub fn tau(&self) -> f32

Get the tau to use when generating text.

pub fn eta(&self) -> f32

Get the eta to use when generating text.

pub fn mu(&self) -> f32

Get the mu to use when generating text.

pub fn repetition_penalty(&self) -> f32

Get the repetition penalty to use when generating text.

pub fn repetition_penalty_range(&self) -> u32

Get the repetition penalty range to use when generating text.

pub fn max_length(&self) -> u32

Get the maximum length to use when generating text.

pub fn stop_on(&self) -> Option<&str>

Get the string to stop on when generating text.

pub fn seed(&self) -> Option<u64>

Get the seed to use when generating text.